feat(ui): replace top control panel with bottom dock and expanded icon rail
Remove the top ControlPanel bar entirely. Add a centered, floating
BottomDock that adapts between live mode (play/pause/reset/zones +
status chips + vehicle count) and replay mode (transport controls +
progress bar + speed buttons). Expand the icon rail from 4 to 7
entries with new Toggles (visibility switches), Speed (sliders),
and Adapter panels. Move the adapter drawer from a right-side panel
to the left panel system.
