+++ apps/ui/src/Controls/BottomDock.module.css
@@ -1,0 +1,229 @@
+.dock {
+  position: absolute;
+  bottom: var(--space-4);
+  left: 50%;
+  transform: translateX(-50%);
+  z-index: 40;
+  display: flex;
+  align-items: center;
+  gap: var(--space-3);
+  height: 44px;
+  padding: 0 var(--space-4);
+  background: rgba(7, 9, 13, 0.88);
+  backdrop-filter: blur(16px);
+  -webkit-backdrop-filter: blur(16px);
+  border: 1px solid var(--color-overlay-1);
+  border-radius: 12px;
+  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.4);
+}
+
+.group {
+  display: flex;
+  align-items: center;
+  gap: 1px;
+  border-radius: var(--radius-sm);
+  overflow: hidden;
+}
+
+.divider {
+  width: 1px;
+  height: 20px;
+  background: var(--color-overlay-1);
+  flex-shrink: 0;
+}
+
+.dockBtn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 32px;
+  height: 32px;
+  border: none;
+  background: var(--surface-bg);
+  cursor: pointer;
+  transition: background var(--transition-fast);
+}
+
+.dockBtn:hover {
+  background: var(--surface-bg-hover);
+}
+
+.dockBtnActive {
+  background: rgba(68, 204, 68, 0.12);
+}
+
+.dockBtnActive:hover {
+  background: rgba(68, 204, 68, 0.2);
+}
+
+.btnIcon {
+  width: 13px;
+  height: 13px;
+  fill: var(--color-gray-light);
+}
+
+.dockBtnActive .btnIcon {
+  fill: #4c4;
+}
+
+/* ── Status chips ── */
+
+.chip {
+  display: inline-flex;
+  align-items: center;
+  gap: var(--space-2);
+  height: 28px;
+  padding: 0 var(--space-3);
+  font-size: var(--text-xs);
+  color: var(--color-gray-light);
+}
+
+.chipActive {
+  color: var(--color-gray-lightest);
+}
+
+.chipLabel {
+  letter-spacing: 0.3px;
+  text-transform: uppercase;
+  opacity: 0.55;
+}
+
+.chipActive .chipLabel {
+  opacity: 0.85;
+}
+
+.led {
+  width: 6px;
+  height: 6px;
+  border-radius: var(--radius-full);
+  background-color: var(--color-gray);
+  flex-shrink: 0;
+}
+
+.ledOn {
+  background-color: #4c4;
+  box-shadow: 0 0 4px #4c4;
+}
+
+/* ── Vehicle count ── */
+
+.vehicleCount {
+  display: inline-flex;
+  align-items: center;
+  gap: var(--space-2);
+  font-variant-numeric: tabular-nums;
+  color: var(--color-gray-lightest);
+  font-size: var(--text-xs);
+  white-space: nowrap;
+}
+
+.vehicleCountValue {
+  font-size: var(--text-base);
+  font-weight: 600;
+  color: var(--color-white);
+}
+
+.vehicleCountLabel {
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+  opacity: 0.55;
+}
+
+/* ── Replay mode ── */
+
+.transportGroup {
+  display: flex;
+  align-items: center;
+  gap: 1px;
+  border-radius: var(--radius-sm);
+  overflow: hidden;
+}
+
+.stopBtn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 32px;
+  height: 32px;
+  border: none;
+  background: var(--surface-bg);
+  cursor: pointer;
+  transition: background var(--transition-fast);
+}
+
+.stopBtn:hover {
+  background: rgba(255, 68, 68, 0.12);
+}
+
+.stopBtn:hover .btnIcon {
+  fill: #f44;
+}
+
+.progressWrap {
+  flex: 0 1 280px;
+  min-width: 80px;
+  height: 6px;
+  background: rgba(255, 255, 255, 0.08);
+  border-radius: 999px;
+  cursor: pointer;
+  position: relative;
+}
+
+.progressFill {
+  height: 100%;
+  background: var(--color-accent, rgba(59, 130, 246, 0.7));
+  border-radius: 999px;
+  transition: width 1s linear;
+  pointer-events: none;
+}
+
+.time {
+  font-size: var(--text-xs);
+  font-variant-numeric: tabular-nums;
+  color: var(--color-gray-light);
+  white-space: nowrap;
+  flex-shrink: 0;
+}
+
+.speedGroup {
+  display: flex;
+  gap: 1px;
+  border-radius: var(--radius-sm);
+  overflow: hidden;
+  flex-shrink: 0;
+}
+
+.speedBtn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 28px;
+  padding: 0 var(--space-3);
+  border: none;
+  background: var(--surface-bg);
+  color: var(--color-gray-light);
+  font-size: var(--text-xs);
+  font-family: inherit;
+  font-weight: 500;
+  cursor: pointer;
+  transition: background var(--transition-fast), color var(--transition-fast);
+}
+
+.speedBtn:hover {
+  background: var(--surface-bg-hover);
+}
+
+.speedBtnActive {
+  background: rgba(57, 153, 255, 0.14);
+  color: var(--color-white);
+}
+
+.fileName {
+  max-width: 120px;
+  font-size: var(--text-xs);
+  color: var(--color-gray);
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  flex-shrink: 0;
+}
