--- apps/ui/src/Controls/Controls.module.css
@@ -1,354 +1,0 @@
-.controlPanel {
-  background: var(--color-bg-dark);
-  display: flex;
-  flex-wrap: nowrap;
-  align-items: center;
-  width: 100%;
-  height: 48px;
-  padding: 0 var(--space-5);
-  gap: var(--space-4);
-  border-bottom: var(--surface-border);
-  font-size: var(--text-sm);
-  color: var(--color-gray-light);
-  box-sizing: border-box;
-  position: relative;
-  overflow: visible;
-}
-
-/* ── Status indicators ── */
-
-.ledDot {
-  width: 6px;
-  height: 6px;
-  border-radius: var(--radius-full);
-  flex-shrink: 0;
-}
-
-.ledOn {
-  background-color: #4c4;
-  box-shadow: 0 0 4px #4c4;
-}
-
-.ledDim {
-  background-color: var(--color-gray);
-}
-
-.ledLabel {
-  font-size: var(--text-xs);
-  letter-spacing: 0.3px;
-  text-transform: uppercase;
-  opacity: 0.35;
-  color: var(--color-gray-light);
-}
-
-.statusChip {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--space-2);
-  height: var(--control-height-sm);
-  padding: 0 var(--space-3);
-  border-radius: var(--radius-sm);
-  background: var(--surface-bg);
-  border: var(--surface-border);
-  white-space: nowrap;
-}
-
-.statusChipActive {
-  background: rgba(57, 153, 255, 0.12);
-  border-color: rgba(57, 153, 255, 0.2);
-}
-
-.statusValue {
-  color: var(--color-gray-lightest);
-  font-size: var(--text-xs);
-}
-
-/* ── Fleet count ── */
-
-.vehicleCount {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--space-3);
-  font-variant-numeric: tabular-nums;
-  color: var(--color-gray-lightest);
-  font-size: var(--text-xs);
-  height: var(--control-height-sm);
-  padding: 0 var(--space-4);
-  border-radius: var(--radius-sm);
-  background: var(--surface-bg);
-  border: var(--surface-border);
-  white-space: nowrap;
-  cursor: pointer;
-  transition:
-    background var(--transition-fast),
-    border-color var(--transition-fast),
-    color var(--transition-fast);
-}
-
-.vehicleCount:hover {
-  background: var(--surface-bg-hover);
-}
-
-.vehicleCountValue {
-  font-size: var(--text-base);
-  font-weight: 600;
-  color: var(--color-white);
-}
-
-.vehicleCountLabel {
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-  opacity: 0.55;
-}
-
-.panelToggleActive {
-  background: rgba(57, 153, 255, 0.12);
-  border-color: rgba(57, 153, 255, 0.24);
-  color: var(--color-white);
-}
-
-.vehicleAlertCount {
-  color: #ffb347;
-}
-
-/* ── Toggles ── */
-
-.toggle {
-  display: flex;
-  align-items: center;
-  gap: var(--space-2);
-  cursor: pointer;
-  user-select: none;
-  height: var(--control-height-sm);
-  padding: 0 var(--space-3);
-  border-radius: var(--radius-sm);
-  transition:
-    background var(--transition-fast),
-    color var(--transition-fast);
-}
-
-.toggle:hover {
-  background: var(--surface-bg);
-}
-
-.toggleActive {
-  background: rgba(57, 153, 255, 0.09);
-}
-
-.toggleLabel {
-  white-space: nowrap;
-  opacity: 0.78;
-}
-
-/* ── Slider options (Speed, Accel, etc.) ── */
-
-.option {
-  position: relative;
-  display: inline-flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: var(--space-3);
-  white-space: nowrap;
-  min-width: 0;
-  height: var(--control-height-sm);
-  padding: 0 var(--space-4);
-  border-radius: var(--radius-sm);
-  background: var(--surface-bg);
-  border: var(--surface-border);
-  overflow: visible;
-}
-
-.firstOption {
-  margin-left: auto;
-}
-
-.optionLabel {
-  flex-shrink: 0;
-  opacity: 0.55;
-  font-size: var(--text-xs);
-  text-transform: uppercase;
-  letter-spacing: 0.6px;
-}
-
-.optionSlider {
-  position: absolute;
-  left: 50%;
-  top: calc(100% + 62px);
-  width: 110px;
-  height: 24px;
-  padding: 0 10px;
-  background: rgba(10, 10, 10, 0.92);
-  border: var(--surface-border);
-  border-radius: 999px;
-  box-shadow: 0 10px 28px rgba(0, 0, 0, 0.45);
-  flex-shrink: 0;
-  -webkit-appearance: none;
-  appearance: none;
-  accent-color: var(--color-accent);
-  transform: translate(-50%, -50%) rotate(-90deg);
-  transform-origin: center;
-  z-index: 2;
-  opacity: 0;
-  pointer-events: none;
-  transition: opacity var(--transition-fast);
-}
-
-.option:hover .optionSlider,
-.option:focus-within .optionSlider {
-  opacity: 1;
-  pointer-events: auto;
-}
-
-.optionSlider::-webkit-slider-runnable-track {
-  height: 3px;
-  background: rgba(255, 255, 255, 0.12);
-  border-radius: 999px;
-}
-
-.optionSlider::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  width: 11px;
-  height: 11px;
-  border-radius: var(--radius-full);
-  background: var(--color-gray-lightest);
-  border: 2px solid rgba(11, 96, 85, 0.8);
-  margin-top: -4px;
-  cursor: pointer;
-  transition: background var(--transition-fast);
-}
-
-.optionSlider::-webkit-slider-thumb:hover {
-  background: var(--color-gray-lighter);
-}
-
-.optionSlider:focus {
-  outline: none;
-}
-
-.optionValue {
-  font-variant-numeric: tabular-nums;
-  min-width: 3ch;
-  margin-left: auto;
-  flex-shrink: 0;
-  font-size: var(--text-xs);
-  opacity: 0.85;
-  color: var(--color-gray-lightest);
-}
-
-/* ── Main buttons (play/pause/reset/zones) ── */
-
-.mainButtonGroup {
-  display: flex;
-  gap: 1px;
-  flex-shrink: 0;
-  overflow: hidden;
-  border-radius: var(--radius-sm);
-}
-
-.mainButton {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: var(--control-height);
-  height: var(--control-height-sm);
-  border: none;
-  background: var(--surface-bg-hover);
-  cursor: pointer;
-  transition: background-color var(--transition-fast);
-}
-
-.mainButton:hover {
-  background: var(--surface-bg-active);
-}
-
-.mainButtonActive {
-  background: rgba(68, 204, 68, 0.12);
-}
-
-.mainButtonActive:hover {
-  background: rgba(68, 204, 68, 0.2);
-}
-
-.mainButtonIcon {
-  width: 13px;
-  height: 13px;
-  fill: var(--color-gray-light);
-}
-
-.mainButtonActive .mainButtonIcon {
-  fill: #4c4;
-}
-
-/* ── Adapter gear button ── */
-
-.gearButton {
-  background: var(--surface-bg);
-  border: var(--surface-border);
-  border-radius: var(--radius-sm);
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  gap: var(--space-3);
-  padding: 0 var(--space-4);
-  height: var(--control-height-sm);
-  color: var(--color-gray-light);
-  transition:
-    opacity var(--transition-fast),
-    background var(--transition-fast);
-}
-
-.gearButton:hover {
-  opacity: 0.9;
-  background: var(--surface-bg-hover);
-}
-
-.gearButton.panelToggleActive {
-  background: rgba(57, 153, 255, 0.12);
-  border-color: rgba(57, 153, 255, 0.24);
-  color: var(--color-white);
-}
-
-.gearButton svg {
-  width: 13px;
-  height: 13px;
-  fill: var(--color-gray-light);
-}
-
-.gearLabel {
-  font-size: var(--text-xs);
-  color: var(--color-gray-lighter);
-}
-
-/* ── Adapter drawer shared styles ── */
-
-.danger {
-  border: 1px solid var(--color-danger-border);
-  border-radius: var(--radius-lg);
-  margin: 0 calc(-1 * var(--space-2));
-  padding: var(--space-4);
-  margin-top: var(--space-3);
-  background: var(--color-danger-bg);
-}
-
-.item {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  margin: var(--space-1) 0;
-}
-
-.itemLabel {
-  font-size: var(--text-base);
-  color: var(--color-gray-lighter);
-  width: 120px;
-}
-
-.itemValue {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  font-size: var(--text-base);
-  color: var(--color-gray-lightest);
-}
