+++ apps/ui/src/Controls/IconRail.module.css
@@ -60,3 +60,7 @@
   line-height: 14px;
   text-align: center;
 }
+
+.spacer {
+  flex: 1;
+}
