+++ apps/ui/src/Controls/IconRail.tsx
@@ -1,8 +1,8 @@
 import classNames from "classnames";
-import { CarIcon, LayersIcon, AlertIcon, RecordCircleIcon } from "@/components/Icons";
+import { CarIcon, LayersIcon, AlertIcon, RecordCircleIcon, EyeIcon, GaugeIcon, Gear } from "@/components/Icons";
 import styles from "./IconRail.module.css";
 
-export type PanelId = "vehicles" | "fleets" | "incidents" | "recordings";
+export type PanelId = "vehicles" | "fleets" | "incidents" | "recordings" | "toggles" | "speed" | "adapter";
 
 interface IconRailProps {
   activePanel: PanelId | null;
@@ -10,34 +10,44 @@ interface IconRailProps {
   incidentCount?: number;
 }
 
-const items: { id: PanelId; Icon: React.FC<React.SVGProps<SVGSVGElement>>; label: string }[] = [
+const topItems: { id: PanelId; Icon: React.FC<React.SVGProps<SVGSVGElement>>; label: string }[] = [
   { id: "vehicles", Icon: CarIcon, label: "Vehicles" },
   { id: "fleets", Icon: LayersIcon, label: "Fleets" },
   { id: "incidents", Icon: AlertIcon, label: "Incidents" },
   { id: "recordings", Icon: RecordCircleIcon, label: "Recordings" },
+  { id: "toggles", Icon: EyeIcon, label: "Visibility" },
+  { id: "speed", Icon: GaugeIcon, label: "Speed" },
+];
+
+const bottomItems: typeof topItems = [
+  { id: "adapter", Icon: Gear, label: "Adapter" },
 ];
 
 export default function IconRail({ activePanel, onPanelChange, incidentCount }: IconRailProps) {
+  const renderButton = ({ id, Icon, label }: (typeof topItems)[number]) => (
+    <button
+      key={id}
+      type="button"
+      className={classNames(styles.railButton, {
+        [styles.railButtonActive]: activePanel === id,
+      })}
+      onClick={() => onPanelChange(activePanel === id ? null : id)}
+      aria-label={label}
+      aria-pressed={activePanel === id}
+      title={label}
+    >
+      <Icon className={styles.railIcon} />
+      {id === "incidents" && incidentCount != null && incidentCount > 0 && (
+        <span className={styles.badge}>{incidentCount > 9 ? "9+" : incidentCount}</span>
+      )}
+    </button>
+  );
+
   return (
     <nav className={styles.rail} aria-label="Sidebar navigation">
-      {items.map(({ id, Icon, label }) => (
-        <button
-          key={id}
-          type="button"
-          className={classNames(styles.railButton, {
-            [styles.railButtonActive]: activePanel === id,
-          })}
-          onClick={() => onPanelChange(activePanel === id ? null : id)}
-          aria-label={label}
-          aria-pressed={activePanel === id}
-          title={label}
-        >
-          <Icon className={styles.railIcon} />
-          {id === "incidents" && incidentCount != null && incidentCount > 0 && (
-            <span className={styles.badge}>{incidentCount > 9 ? "9+" : incidentCount}</span>
-          )}
-        </button>
-      ))}
+      {topItems.map(renderButton)}
+      <div className={styles.spacer} />
+      {bottomItems.map(renderButton)}
     </nav>
   );
 }
