+++ apps/ui/src/Controls/SpeedPanel.module.css
@@ -1,0 +1,19 @@
+.header {
+  padding: var(--space-4) var(--space-5);
+  border-bottom: 1px solid var(--color-overlay-1);
+  flex-shrink: 0;
+}
+
+.title {
+  font-size: var(--text-sm);
+  font-weight: 600;
+  color: var(--color-text-primary);
+  margin: 0;
+}
+
+.body {
+  padding: var(--space-4) var(--space-5);
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-4);
+}
