--- apps/ui/src/Map/ReplayBar.module.css
@@ -1,121 +1,0 @@
-.bar {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  z-index: 40;
-  display: flex;
-  align-items: center;
-  gap: var(--space-3);
-  height: 40px;
-  padding: 0 var(--space-4);
-  background: rgba(7, 9, 13, 0.88);
-  backdrop-filter: blur(12px);
-  -webkit-backdrop-filter: blur(12px);
-  border-top: 1px solid var(--color-overlay-1);
-}
-
-.fileName {
-  max-width: 150px;
-  font-size: var(--text-xs);
-  color: var(--color-gray);
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  flex-shrink: 0;
-}
-
-.transportBtn {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 28px;
-  height: 28px;
-  border: none;
-  background: var(--surface-bg);
-  border-radius: var(--radius-sm);
-  cursor: pointer;
-  flex-shrink: 0;
-  transition: background var(--transition-fast);
-}
-
-.transportBtn:hover {
-  background: var(--surface-bg-hover);
-}
-
-.stopBtn {
-  composes: transportBtn;
-}
-
-.stopBtn:hover {
-  background: rgba(255, 68, 68, 0.12);
-}
-
-.transportIcon {
-  width: 12px;
-  height: 12px;
-  fill: var(--color-gray-light);
-}
-
-.stopBtn:hover .transportIcon {
-  fill: #f44;
-}
-
-.progressWrap {
-  flex: 1;
-  height: 6px;
-  background: rgba(255, 255, 255, 0.08);
-  border-radius: 999px;
-  cursor: pointer;
-  position: relative;
-  min-width: 60px;
-}
-
-.progressFill {
-  height: 100%;
-  background: var(--color-accent, rgba(59, 130, 246, 0.7));
-  border-radius: 999px;
-  transition: width 1s linear;
-  pointer-events: none;
-}
-
-.time {
-  font-size: var(--text-xs);
-  font-variant-numeric: tabular-nums;
-  color: var(--color-gray-light);
-  white-space: nowrap;
-  flex-shrink: 0;
-}
-
-.speedGroup {
-  display: flex;
-  gap: 1px;
-  border-radius: var(--radius-sm);
-  overflow: hidden;
-  flex-shrink: 0;
-}
-
-.speedBtn {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 28px;
-  padding: 0 var(--space-3);
-  border: none;
-  background: var(--surface-bg);
-  color: var(--color-gray-light);
-  font-size: var(--text-xs);
-  font-family: inherit;
-  font-weight: 500;
-  cursor: pointer;
-  transition: background var(--transition-fast), color var(--transition-fast);
-}
-
-.speedBtn:hover {
-  background: var(--surface-bg-hover);
-}
-
-.speedBtnActive {
-  background: rgba(57, 153, 255, 0.14);
-  color: var(--color-white);
-}
