+++ apps/ui/src/components/Icons.tsx
@@ -265,3 +265,22 @@ export function RecordCircleIcon(props: SVGProps) {
     </svg>
   );
 }
+
+export function EyeIcon(props: SVGProps) {
+  return (
+    <svg viewBox="0 0 24 24" {...props}>
+      <path d="M12 4.5C7 4.5 2.73 7.61 1 12c1.73 4.39 6 7.5 11 7.5s9.27-3.11 11-7.5c-1.73-4.39-6-7.5-11-7.5zm0 12.5c-2.76 0-5-2.24-5-5s2.24-5 5-5 5 2.24 5 5-2.24 5-5 5zm0-8c-1.66 0-3 1.34-3 3s1.34 3 3 3 3-1.34 3-3-1.34-3-3-3z" />
+    </svg>
+  );
+}
+
+export function GaugeIcon(props: SVGProps) {
+  return (
+    <svg viewBox="0 0 24 24" {...props}>
+      <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.42 0-8-3.58-8-8s3.58-8 8-8 8 3.58 8 8-3.58 8-8 8z" />
+      <path d="M12 6v2m-4.24.76l1.42 1.42M6 12h2m.76 4.24l1.42-1.42M12 16v2m4.24-.76l-1.42-1.42M18 12h-2m-.76-4.24l-1.42 1.42" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" fill="none" />
+      <circle cx="12" cy="12" r="1.5" />
+      <path d="M12 12l3-5" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" fill="none" />
+    </svg>
+  );
+}
